backend/pyproject.toml
[project]
name = "erp-backend"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2",
 "environs[django]>=14.1.1",
 "strawberry-graphql-django>=0.58.0",
]

[dependency-groups]
dev = ["django-stubs>=5.1.3", "ruff>=0.11.4"]

[tool.pyright]
typeCheckingMode = 'basic'
venvPath = '.'
venv = './.venv'
exclude = ['./.venv']
